+++ Source/DerivedQuantities/DerivedQuantity.h
@@ -4,7 +4,7 @@
  * @date 6/06/2013
  * @section LICENSE
  *
- * Copyright NIWA Science ©2013 - www.niwa.co.nz
+ * Copyright NIWA Science ï¿½2013 - www.niwa.co.nz
  *
  * @section DESCRIPTION
  *
@@ -36,13 +36,13 @@ public:
   void                        Validate();
   void                        Build();
   void                        Reset();
-  float                       GetValue(unsigned year);
-  float                       GetValue(unsigned year, unsigned row, unsigned col);
-  float                       GetInitialisationValue(unsigned phase = 0, unsigned index = 0);
-  float                       GetInitialisationValue(unsigned row, unsigned col, unsigned phase, unsigned index);
+  double                      GetValue(unsigned year);
+  double                      GetValue(unsigned year, unsigned row, unsigned col);
+  double                      GetInitialisationValue(unsigned phase = 0, unsigned index = 0);
+  double                      GetInitialisationValue(unsigned row, unsigned col, unsigned phase, unsigned index);
 
-  float                       GetLastValueFromInitialisation(unsigned phase);
-  float                       GetLastValueFromInitialisation(unsigned phase, unsigned row, unsigned col);
+  double                      GetLastValueFromInitialisation(unsigned phase);
+  double                      GetLastValueFromInitialisation(unsigned phase, unsigned row, unsigned col);
 
   // pure methods
   virtual void                DoValidate() = 0;
@@ -50,8 +50,8 @@ public:
 
   // accessors
   const string&               time_step() { return time_step_label_; }
-  vector<vector<float> >&     initialisation_values() { return initialisation_values_; }
-  const map<unsigned, float>& values() { return values_; }
+  vector<vector<double> >&     initialisation_values() { return initialisation_values_; }
+  const map<unsigned, double>& values() { return values_; }
   bool                        is_spatial() {return spatial_;};
 protected:
   // Members
@@ -59,20 +59,20 @@ protected:
   WorldView*                  world_ = nullptr;
   string                      time_step_label_ = "";
   unsigned                    current_initialisation_phase_ = 0;
-  vector<vector<float>>       initialisation_values_;
-  vector<vector<vector<vector<float>>>>  initialisation_values_by_space_;  //[phase][row][col][value]
+  vector<vector<double>>       initialisation_values_;
+  vector<vector<vector<vector<double>>>>  initialisation_values_by_space_;  //[phase][row][col][value]
 
-  map<unsigned, float>        values_;
-  map<unsigned, vector<vector<float>>> values_by_space_;  // year x row x col
-  float                       cache_value_;
-  float                       value_;
+  map<unsigned, double>        values_;
+  map<unsigned, vector<vector<double>>> values_by_space_;  // year x row x col
+  double                       cache_value_;
+  double                       value_;
 
   string                      proportion_method_;
-  float                       time_step_proportion_;
+  double                       time_step_proportion_;
   bool                        spatial_ = false;
 
   // objects for thread safety of rng
-  vector<float>                       random_numbers_;
+  vector<double>                       random_numbers_;
   unsigned                            n_agents_;
   vector<vector<unsigned>>            cell_offset_;
